requirements.txt
fastapi
matplotlib
numpy
openai
pillow
pydantic
python-dotenv
stability-sdk
uvicorn